staff list works, vertical formatting is better
