+++ app/(tabs)/incident_command.tsx
@@ -7,13 +7,14 @@ import ParallaxScrollView from '@/components/ParallaxScrollView';
 import { IconSymbol } from '@/components/ui/IconSymbol';
 import { Patient, ColorScheme } from '@/components/ic_dash/types';
 import  PatientTable  from '@/components/ic_dash/pt_table';
-import  { priorityColors, initialPatients }  from '@/components/ic_dash/constants'
+import  StaffList  from '@/components/ic_dash/staff_list';
+
+import  { priorityColors, initialPatients, zoneStaff }  from '@/components/ic_dash/constants'
 
 
 
 export default function IncidentCommandDashboard() {
   const [patients, setPatients] =  useState<Patient[]>(initialPatients);
-  const [openZone, setOpenZone] = useState<string | null>(null);
 
   // Calculate patient counts by priority
   const priorityCounts = patients.reduce<Record<string, number>>((counts, patient) => {
@@ -21,14 +22,6 @@ export default function IncidentCommandDashboard() {
     counts[priority] = (counts[priority] || 0) + 1;
     return counts;
   }, {});
-
-
-  const zoneStaff: Record<string, string []>= {
-    'Zone 1': ['person1'],
-    'Zone 2': ['person2','person3','person4'],
-    'Zone 3': ['person5','person6'],
-    'Zone 4': ['person7', 'person8', 'person9', 'person10'],
-  };
 
 
   // Render priority status card
@@ -39,25 +32,21 @@ export default function IncidentCommandDashboard() {
     </View>
   );
 
-   // Toggle function to expand or collapse the zone list
-   const toggleZone = (zone: string) => {
-    setOpenZone(openZone === zone ? null : zone); // Toggle the clicked zone's state
-  };
-
   return (
     <ThemedView style={styles.container}>
       <ThemedView style={styles.header}>
-        <ThemedText style={styles.title}>TriageNow</ThemedText>
-        <ThemedText style={styles.subtitle}>Incident Command Dashboard</ThemedText>
+        <ThemedView style={styles.header}>
+          <ThemedText style={styles.title}>TriageNow</ThemedText>
+          <ThemedText style={styles.subtitle}>Incident Command Dashboard</ThemedText>
+        </ThemedView>
+
+        <ScrollView horizontal style={styles.priorityCardsContainer}>
+          {renderPriorityCard('IMMEDIATE', priorityCounts['IMMEDIATE'], priorityColors.IMMEDIATE)}
+          {renderPriorityCard('DELAYED', priorityCounts['DELAYED'], priorityColors.DELAYED)}
+          {renderPriorityCard('MINOR', priorityCounts['MINOR'], priorityColors.MINOR)}
+          {renderPriorityCard('EXPECTANT', priorityCounts['EXPECTANT'], priorityColors.EXPECTANT)}
+        </ScrollView>
       </ThemedView>
-
-      <ScrollView horizontal style={styles.priorityCardsContainer}>
-        {renderPriorityCard('IMMEDIATE', priorityCounts['IMMEDIATE'], priorityColors.IMMEDIATE)}
-        {renderPriorityCard('DELAYED', priorityCounts['DELAYED'], priorityColors.DELAYED)}
-        {renderPriorityCard('MINOR', priorityCounts['MINOR'], priorityColors.MINOR)}
-        {renderPriorityCard('EXPECTANT', priorityCounts['EXPECTANT'], priorityColors.EXPECTANT)}
-      </ScrollView>
-
       <ThemedView style={styles.mainBody}>
 
         {/* patient table in left column */}
@@ -67,27 +56,8 @@ export default function IncidentCommandDashboard() {
 
         {/* staff list by zone in right column */}
         <ThemedView style={styles.staffList}>
-          {Object.entries(zoneStaff).map(([zone, staff]) => (
-          <View key={zone} style={styles.zoneContainer}>
-            <TouchableOpacity
-              style={styles.zoneTitle}
-              onPress={() => toggleZone(zone)} // Toggle visibility of the staff list
-            >
-              <Text style={styles.zoneText}>{zone}</Text>
-            </TouchableOpacity>
+          <StaffList zoneStaff={zoneStaff} />       
 
-            {/* Render staff list if this zone is open */}
-            {openZone === zone && (
-              <View style={styles.staffList}>
-                {staff.map((person, index) => (
-                  <Text key={index} style={styles.staffText}>
-                    {person}
-                  </Text>
-                ))}
-              </View>
-            )}
-          </View>
-        ))}
         </ThemedView>
       </ThemedView>
     </ThemedView>
@@ -110,7 +80,7 @@ const styles = StyleSheet.create({
   subtitle: {
     fontSize: 28,
     fontWeight: 'bold',
-    marginBottom: 16,
+    marginBottom: 0,
   },
   priorityCardsContainer: {
     flexDirection: 'row',
@@ -148,29 +118,11 @@ const styles = StyleSheet.create({
     borderRadius: 8,
     overflow: 'hidden',
   },
-
-  stafflist: {
-    backgroundColor: 'e8e8e8',
-    flex: 1,
-  },
-  zoneContainer: {
-    marginBottom: 20,
-  },
-  zoneTitle: {
-    padding: 10,
-    backgroundColor: '#f1f1f1',
-    borderRadius: 5,
-    marginBottom: 5,
-  },
-  zoneText: {
-    fontSize: 18,
-    fontWeight: 'bold',
-  },
   staffList: {
+    backgroundColor: 'e8e8e8',
+    flex: 0.5,
     marginLeft: 20,
     paddingTop: 5,
   },
-  staffText: {
-    fontSize: 16,
-  },
+
 });
